major cleanup and reorganization
