+++ mlpf/pyg/README.md
@@ -13,6 +13,7 @@ The dataset is available from zenodo: https://doi.org/10.5281/zenodo.4452283.
 
 To download and process the full DELPHES dataset:
 ```bash
+cd delphes/
 ./get_data_delphes.sh
 ```
 
@@ -21,7 +22,7 @@ This script will download and process the data under a directory called `data/de
 To perform a quick training on the dataset:
 ```bash
 cd ../
-python -u pyg_pipeline.py --data delphes --dataset=<path_to_delphes_data> --dataset_qcd=<path_to_delphes_data>
+python -u pyg_pipeline.py --data delphes --dataset=../data/delphes/ --dataset_test=../data/delphes/pythia8_ttbar
 ```
 
 To load a pretrained model which is stored in a directory under `particleflow/experiments` for evaluation:
@@ -34,6 +35,7 @@ python -u pyg_pipeline.py --data delphes --load --load_model=<model_directory> -
 
 To download and process the full CMS dataset:
 ```bash
+cd cms/
 ./get_data_cms.sh
 ```
 This script will download and process the data under a directory called `data/cms` under `particleflow`.
@@ -41,7 +43,28 @@ This script will download and process the data under a directory called `data/cm
 To perform a quick training on the dataset:
 ```bash
 cd ../
-python -u pyg_pipeline.py --data cms --dataset=<path_to_cms_data> --dataset_qcd=<path_to_cms_data>
+python -u pyg_pipeline.py --data cms --dataset=../data/cms/TTbar_14TeV_TuneCUETP8M1_cfi/ --dataset_test=../data/cms/QCDForPF_14TeV_TuneCUETP8M1_cfi/
+```
+
+To load a pretrained model which is stored in a directory under `particleflow/experiments` for evaluation:
+```bash
+cd ../
+python -u pyg_pipeline.py --data cms --load --load_model=<model_directory> --load_epoch=<epoch_to_load> --dataset=<path_to_cms_data> --dataset_qcd=<path_to_cms_data>
+```
+
+### CLIC training
+
+To download and process the full CLIC dataset:
+```bash
+cd clic/
+./get_data_clic.sh
+```
+This script will download and process the data under a directory called `data/clic` under `particleflow`.
+
+To perform a quick training on the dataset:
+```bash
+cd ../
+python -u pyg_pipeline.py --data clic --dataset=../data/gev380ee_pythia6_ttbar_rfull201/ --dataset_test=../data/gev380ee_pythia6_qcd_all_rfull201/
 ```
 
 To load a pretrained model which is stored in a directory under `particleflow/experiments` for evaluation:
