+++ mlpf/pyg/cms/cms_utils.py
@@ -28,14 +28,14 @@ ELEM_NAMES_CMS = [
 # https://github.com/cms-sw/cmssw/blob/master/DataFormats/ParticleFlowCandidate/src/PFCandidate.cc#L254
 CLASS_LABELS_CMS = [0, 211, 130, 1, 2, 22, 11, 13, 15]
 CLASS_NAMES_CMS_LATEX = [
-    r"none",
-    r"chhad",
-    r"nhad",
-    r"HFEM",
-    r"HFHAD",
-    r"$\gamma$",
-    r"$e^\pm$",
-    r"$\mu^\pm$",
+    "none",
+    "chhad",
+    "nhad",
+    "HFEM",
+    "HFHAD",
+    "$\gamma$",
+    "$e^\pm$",
+    "$\mu^\pm$",
     r"$\tau$",
 ]
 CLASS_NAMES_CMS = [
@@ -71,7 +71,7 @@ CMS_PF_CLASS_NAMES = [
     "muon",
 ]
 
-X_FEATURES = [
+X_FEATURES_CMS = [
     "typ_idx",
     "pt",
     "eta",
@@ -116,7 +116,6 @@ X_FEATURES = [
 ]
 
 Y_FEATURES = [
-    "typ_idx",
     "charge",
     "pt",
     "eta",
@@ -127,13 +126,19 @@ Y_FEATURES = [
 
 
 def prepare_data_cms(fn):
+    """
+    Takes as input a bz2 file that contains the cms raw information, and returns a list of PyG Data() objects.
+    Each element of the list looks like this ~ Data(x=[#, 41], ygen=[#, 6], ygen_id=[#, 9], ycand=[#, 6], ycand_id=[#, 9])
+
+    Args
+        raw_file_name: raw parquet data file.
+    Returns
+        list of Data() objects.
+    """
 
     batched_data = []
 
-    if fn.endswith(".pkl"):
-        data = pickle.load(open(fn, "rb"), encoding="iso-8859-1")
-    elif fn.endswith(".pkl.bz2"):
-        data = pickle.load(bz2.BZ2File(fn, "rb"))
+    data = pickle.load(bz2.BZ2File(fn, "rb"))
 
     for event in data:
         Xelem = event["Xelem"]
@@ -151,8 +156,7 @@ def prepare_data_cms(fn):
             Xelem,
             "typ_idx",
             np.array(
-                [ELEM_LABELS_CMS.index(int(i)) for i in Xelem["typ"]],
-                dtype=np.float32,
+                [ELEM_LABELS_CMS.index(int(i)) for i in Xelem["typ"]], dtype=np.float32
             ),
         )
         ygen = append_fields(
@@ -173,7 +177,7 @@ def prepare_data_cms(fn):
         )
 
         Xelem_flat = np.stack(
-            [Xelem[k].view(np.float32).data for k in X_FEATURES],
+            [Xelem[k].view(np.float32).data for k in X_FEATURES_CMS],
             axis=-1,
         )
         ygen_flat = np.stack(
