+++ mlpf/pyg/cms/get_data_cms.sh
@@ -2,9 +2,6 @@
 
 set -e
 
-# make data/ directory to hold the cms/ directory of datafiles under particleflow/
-mkdir -p ../../data
-
 # get the cms data
 rsync -r --progress lxplus.cern.ch:/eos/user/j/jpata/mlpf/cms .
 
@@ -17,12 +14,19 @@ for sample in cms/* ; do
     cd ../../../
 done
 
-# process the cms data
-for sample in cms/* ; do
+# make data/ directory to hold the cms/ directory of datafiles under particleflow/
+mkdir -p ../../../data
+
+# move the cms/ directory of datafiles there
+mv cms ../../../data/
+cd ..
+
+# process the raw datafiles
+echo -----------------------
+for sample in ../../data/cms/* ; do
   echo $sample
   #generate pytorch data files from pkl files
   python3 PFGraphDataset.py --data cms --dataset $sample \
     --processed_dir $sample/processed --num-files-merge 1 --num-proc 1
 done
-
-mv cms ../../data/
+echo -----------------------
