+++ mlpf/pyg/delphes/get_data_delphes.sh
@@ -2,9 +2,6 @@
 
 set -e
 
-# make data/ directory to hold the delphes/ directory of datafiles under particleflow/
-mkdir -p ../../data
-
 # make delphes directories
 mkdir -p delphes/pythia8_ttbar/raw
 mkdir -p delphes/pythia8_ttbar/processed
@@ -30,16 +27,19 @@ do
     wget --no-check-certificate -nc https://zenodo.org/record/4559324/files/tev14_pythia8_qcd_10_"$i".pkl.bz2
 done
 bzip2 -d *
+cd ../../../
 
-# get back in the pytorch directory
-cd ../../../../
+# make data/ directory to hold the delphes/ directory of datafiles under particleflow/
+mkdir -p ../../../data
+
+# move the delphes/ directory of datafiles there
+mv delphes ../../../data/
+cd ..
 
 #generate pytorch data files from pkl files
-python3 PFGraphDataset.py --data delphes --dataset delphes/pythia8_ttbar \
-  --processed_dir delphes/pythia8_ttbar/processed --num-files-merge 1 --num-proc 1
+python3 PFGraphDataset.py --data delphes --dataset ../../data/delphes/pythia8_ttbar \
+  --processed_dir ../../data/delphes/pythia8_ttbar/processed --num-files-merge 1 --num-proc 1
 
 #generate pytorch data files from pkl files
-python3 PFGraphDataset.py --data delphes --dataset delphes/pythia8_qcd \
-  --processed_dir delphes/pythia8_qcd/processed --num-files-merge 1 --num-proc 1
-
-mv delphes ../../data/
+python3 PFGraphDataset.py --data delphes --dataset ../../data/delphes/pythia8_qcd \
+  --processed_dir ../../data/delphes/pythia8_qcd/processed --num-files-merge 1 --num-proc 1
